assistant_api/tests/functional/requirements.txt
async-fastapi-jwt-auth==0.6.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiohttp==3.9.3
pytest==7.4.4
pytest-asyncio==0.12.0
pydantic==2.3.0
starlette==0.27.0
async-fastapi-jwt-auth==0.6.2
requests==2.31.0
